refactor: modify admin pages
