+++ packages/gdsc-dju-admin/src/components/common/AdminSidebar/index.tsx
@@ -2,7 +2,7 @@ import React, { memo } from 'react';
 import { useLocation, useNavigate } from 'react-router';
 import { useSearchParams } from 'react-router-dom';
 
-import { SidebarContainer } from '../../../pages/Admin/styled';
+import { SidebarContainer } from '../../../pages/styled';
 import { ROUTES } from '../../../routes/Route';
 
 import {
