+++ packages/gdsc-dju-admin/src/pages/AdminApplicants/styled.ts
