+++ packages/gdsc-dju-admin/src/pages/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { Route, Routes } from 'react-router-dom';
-import AdminSidebar from '../../components/common/AdminSidebar';
+import AdminSidebar from '../components/common/AdminSidebar';
 
 import AdminApplicants from './AdminApplicants';
 import AdminEmail from './AdminEmail';
